Move tt code to dedicated spculative decoding folder
