+++ models/demos/deepseek_v3_d_p/tt/speculative_decoding/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Tenstorrent device modules for speculative-decoding drafters"""
